Handle arch specification in Julia script
Currently the Julia script assumes that the architecture for Linux is
x86-64. This adjusts the logic such that `config[:arch]` is handled
appropriately for arm64, which Julia calls AArch64.

Note that not all versions of Julia have AArch64 binaries available, as
it's not considered a tier 1 platform (so binaries are usually but not
always available for releases). Thus requesting a particular Julia
version with `arch: arm64` may fail, probably with some kind of weird,
cryptic message from curl. The case of nightly binaries with ARM is
handled explicitly here though, as we do not now nor have we ever had
ARM nightly binaries.